--- src/main/java/com/goormitrip/config/SecurityConfig.java
@@ -1,43 +1,0 @@
-package com.goormitrip.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.provisioning.InMemoryUserDetailsManager;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    @Bean
-    public UserDetailsService userDetailsService() {
-        UserDetails user = User
-            .withUsername("user")
-            .password("{noop}goormi123@") 
-            .roles("USER")
-            .build();
-
-        return new InMemoryUserDetailsManager(user);
-    }
-    
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-            .csrf(csrf -> csrf.disable())
-            .authorizeHttpRequests(auth -> auth
-                .requestMatchers("/api/products/**").permitAll()
-                .anyRequest().authenticated()                 
-            )
-            .formLogin(form -> form.disable())
-            .httpBasic(Customizer.withDefaults());      
-
-        return http.build();
-    }
-
-}
