--- src/main/java/com/goormitrip/domain/Product.java
@@ -1,41 +1,0 @@
-package com.goormitrip.domain;
-
-import jakarta.persistence.*;
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-public class Product {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    
-    private String title;
-
-    @Column
-    private String description;
-
-    private BigDecimal price;
-
-    private LocalDateTime startDate;
-
-    private LocalDateTime endDate;
-
-    private String region;
-
-    private String theme;
-
-    @Enumerated(EnumType.STRING)
-    private ProductStatus status;
-
-    private LocalDateTime createdAt;
-    
-    private LocalDateTime updatedAt;
-}
