--- src/main/java/com/goormitrip/domain/ProductStatus.java
@@ -1,7 +1,0 @@
-package com.goormitrip.domain;
-
-public enum ProductStatus {
-    ACTIVE,
-    CLOSED,
-    HIDDEN
-}
