--- src/main/java/com/goormitrip/domain/Reservation.java
@@ -1,41 +1,0 @@
-package com.goormitrip.domain;
-
-import jakarta.persistence.*;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-import com.goormitrip.domain.Product;
-//import com.goormitrip.domain.User;
-import com.goormitrip.domain.ReservationStatus;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table(name = "reservation")
-public class Reservation {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    //@ManyToOne(fetch = FetchType.LAZY)
-    //@JoinColumn(name = "user_id", nullable = false)
-    //private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "product_id", nullable = false)
-    private Product product;
-
-    private LocalDate travelDate;
-
-    private int peopleCount;
-
-    @Enumerated(EnumType.STRING)
-    private ReservationStatus status;
-
-    private LocalDateTime createdAt;
-
-    private LocalDateTime updatedAt;
-}
