--- src/main/java/com/goormitrip/domain/ReservationStatus.java
@@ -1,8 +1,0 @@
-package com.goormitrip.domain;
-
-public enum ReservationStatus {
-    RESERVED,
-    CANCELLED,
-    PENDING,
-    COMPLETED
-}
