--- src/main/java/com/goormitrip/dto/CompareRequest.java
@@ -1,9 +1,0 @@
-package com.goormitrip.dto;
-
-import java.util.List;
-import lombok.Data;
-
-@Data
-public class CompareRequest {
-    private List<Long> ids;
-}
