--- src/main/java/com/goormitrip/goormitrip/config/SecurityConfig.java
@@ -1,71 +1,0 @@
-    package com.goormitrip.goormitrip.config;
-
-    import org.springframework.context.annotation.Bean;
-    import org.springframework.context.annotation.Configuration;
-    import org.springframework.security.authentication.AuthenticationManager;
-    import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-    import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-    import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-    import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-    import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-    import org.springframework.security.config.http.SessionCreationPolicy;
-    import org.springframework.security.core.userdetails.UserDetailsService;
-    import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-    import org.springframework.security.crypto.password.PasswordEncoder;
-    import org.springframework.security.web.SecurityFilterChain;
-    import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-    import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
-
-    import com.goormitrip.goormitrip.security.JwtAuthFilter;
-
-    @Configuration
-    @EnableWebSecurity
-    public class SecurityConfig {
-
-        private final UserDetailsService userDetailsService;
-
-        private final JwtAuthFilter jwtAuthFilter;
-
-        public SecurityConfig(UserDetailsService userDetailsService, JwtAuthFilter jwtAuthFilter) {
-            this.userDetailsService = userDetailsService;
-            this.jwtAuthFilter = jwtAuthFilter;
-        }
-
-        @Bean
-        public PasswordEncoder passwordEncoder(){
-            return new BCryptPasswordEncoder();
-        }
-
-        @Bean
-        public AuthenticationManager authenticationManager(
-            AuthenticationConfiguration authenticationConfiguration) throws Exception {
-            return authenticationConfiguration.getAuthenticationManager();
-        }
-
-        @Bean
-        public DaoAuthenticationProvider authenticationProvider() {
-            DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-            authProvider.setUserDetailsService(userDetailsService);
-            authProvider.setPasswordEncoder(passwordEncoder());
-
-            return authProvider;
-        }
-
-        @Bean
-        public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-            http.csrf(AbstractHttpConfigurer::disable)
-                .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-                .authorizeHttpRequests(auth ->
-                    auth.requestMatchers(
-                        new AntPathRequestMatcher("/"),
-                        new AntPathRequestMatcher("/users/**"),
-                        new AntPathRequestMatcher("/auth/**"),
-                        new AntPathRequestMatcher("/error")
-                    ).permitAll().anyRequest().authenticated()
-                )
-                .authenticationProvider(authenticationProvider())
-                .addFilterBefore(jwtAuthFilter, UsernamePasswordAuthenticationFilter.class);
-
-            return http.build();
-        }
-    }
