--- src/main/java/com/goormitrip/goormitrip/domain/UserEntity.java
@@ -1,76 +1,0 @@
-package com.goormitrip.goormitrip.domain;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-
-import org.hibernate.annotations.CreationTimestamp;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.time.LocalDateTime;
-import java.util.*;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "users")
-
-public class UserEntity implements UserDetails {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column(unique = true, nullable = false)
-	private String email;
-
-	@Column(nullable = false)
-	private String password;
-	private String phone;
-
-	@Enumerated(EnumType.STRING)
-	@Column(nullable = false)
-	private UserRole role = UserRole.USER;
-
-	@CreationTimestamp
-	private LocalDateTime createdAt;
-
-	public enum UserRole {
-		USER, ADMIN
-	}
-	
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		return Collections.singletonList(new SimpleGrantedAuthority(("ROLE_" + role.name())));
-	}
-
-	@Override
-	public String getUsername() {
-		return email;
-	}
-
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return true;
-	}
-
-}
-
-
