--- src/main/java/com/goormitrip/goormitrip/dto/AuthResponse.java
@@ -1,15 +1,0 @@
-package com.goormitrip.goormitrip.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthResponse {
-
-	private String token;
-	private String email;
-	private String role;
-}
