+++ src/main/java/com/goormitrip/goormitrip/global/security/JwtUtils.java
@@ -1,0 +1,83 @@
+package com.goormitrip.goormitrip.global.security;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.security.Keys;
+
+import javax.crypto.SecretKey;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Component;
+
+import jakarta.annotation.PostConstruct;
+import java.util.*;
+import java.util.function.Function;
+
+@Component
+public class JwtUtils {
+
+	private SecretKey key;
+
+	@Value("${app.jwt.secret}")
+	private String jwtSecret;
+
+	@Value("${app.jwt.expiration}")
+	private long jwtExpiration;
+
+	@PostConstruct
+	public void init() {
+		byte[] keyBytes = jwtSecret.getBytes();
+		this.key = Keys.hmacShaKeyFor(keyBytes);
+	}
+
+	public String extractUsername(String token) {
+		return extractClaim(token, Claims::getSubject);
+	}
+
+	public Date extractExpiration(String token) {
+		return extractClaim(token, Claims::getExpiration);
+	}
+
+	public <T> T extractClaim(String token, Function<Claims, T> claimsResolver) {
+		final Claims claims = extractAllClaims(token);
+		return claimsResolver.apply(claims);
+	}
+
+	private Claims extractAllClaims(String token) {
+		try {
+			return Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token).getBody();
+		} catch (io.jsonwebtoken.ExpiredJwtException e) {
+			return e.getClaims();
+		}
+	}
+
+	private Boolean isTokenExpired(String token) {
+		try {
+			return extractExpiration(token).before(new Date());
+		} catch (io.jsonwebtoken.ExpiredJwtException e) {
+			return true;
+		}
+	}
+
+	public String generateToken(UserDetails userDetails) {
+		Map<String, Object> claims = new HashMap<>();
+		return createToken(claims, userDetails.getUsername());
+	}
+
+	private String createToken(Map<String, Object> claims, String subject) {
+		return Jwts.builder()
+			.setClaims(claims)
+			.setSubject(subject)
+			.setIssuedAt(new Date(System.currentTimeMillis()))
+			.setExpiration(new Date(System.currentTimeMillis() + jwtExpiration))
+			.signWith(key, SignatureAlgorithm.HS512)
+			.compact();
+	}
+
+	public boolean isTokenValid(String token, UserDetails userDetails) {
+		final String username = extractUsername(token);
+		return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
+	}
+}
