+++ src/main/java/com/goormitrip/goormitrip/product/domain/ProductStatus.java
@@ -1,0 +1,7 @@
+package com.goormitrip.goormitrip.product.domain;
+
+public enum ProductStatus {
+    ACTIVE,
+    CLOSED,
+    HIDDEN
+}
