+++ src/main/java/com/goormitrip/goormitrip/product/domain/ReservationStatus.java
@@ -1,0 +1,8 @@
+package com.goormitrip.goormitrip.product.domain;
+
+public enum ReservationStatus {
+    RESERVED,
+    CANCELLED,
+    PENDING,
+    COMPLETED
+}
