+++ src/main/java/com/goormitrip/goormitrip/product/dto/CompareRequest.java
@@ -1,0 +1,9 @@
+package com.goormitrip.goormitrip.product.dto;
+
+import java.util.List;
+import lombok.Data;
+
+@Data
+public class CompareRequest {
+    private List<Long> ids;
+}
