+++ src/main/java/com/goormitrip/goormitrip/product/service/ReservationService.java
@@ -1,0 +1,8 @@
+package com.goormitrip.goormitrip.product.service;
+
+import java.time.LocalDate;
+import java.util.List;
+
+public interface ReservationService {
+    List<LocalDate> getAvailableDates(Long productId);
+}
