--- src/main/java/com/goormitrip/goormitrip/security/JwtAuthFilter.java
@@ -1,68 +1,0 @@
-package com.goormitrip.goormitrip.security;
-
-import java.io.IOException;
-
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-
-@Component
-@RequiredArgsConstructor
-public class JwtAuthFilter extends OncePerRequestFilter {
-
-	private final JwtUtils jwtUtils;
-
-	private final UserDetailsService userDetailsService;
-
-	@Override
-	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
-		FilterChain filterChain) throws ServletException, IOException {
-
-		final String authHeader = request.getHeader("Authorization");
-		final String jwt;
-		final String userEmail;
-
-		if (authHeader == null || !authHeader.startsWith("Bearer ")) {
-			filterChain.doFilter(request, response);
-			return;
-		}
-
-		try {
-			jwt = authHeader.substring(7).trim();
-			userEmail = jwtUtils.extractUsername(jwt);
-
-			if (userEmail != null && SecurityContextHolder.getContext().getAuthentication() == null) {
-				UserDetails userDetails = this.userDetailsService.loadUserByUsername((userEmail));
-
-				if (jwtUtils.isTokenValid(jwt, userDetails)) {
-					UsernamePasswordAuthenticationToken authToken = new UsernamePasswordAuthenticationToken(
-						userDetails,
-						null,
-						userDetails.getAuthorities()
-					);
-					authToken.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
-					SecurityContextHolder.getContext().setAuthentication(authToken);
-				} else {
-					response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Invalid JWT token");
-					return;
-				}
-			}
-		} catch (Exception e) {
-			response.sendError(HttpServletResponse.SC_UNAUTHORIZED, "Authentication failed");
-			return;
-		}
-
-		filterChain.doFilter(request, response);
-
-	}
-}
