--- src/main/java/com/goormitrip/goormitrip/service/UserDetailsServiceImpl.java
@@ -1,25 +1,0 @@
-package com.goormitrip.goormitrip.service;
-
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import com.goormitrip.goormitrip.repository.UserRepository;
-
-@Service
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-	private final UserRepository userRepository;
-
-	public UserDetailsServiceImpl(UserRepository userRepository) {
-		this.userRepository = userRepository;
-	}
-
-	@Override
-	public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-		return userRepository.findByEmail(email)
-			.orElseThrow(() -> new UsernameNotFoundException("이메일 또는 비밀번호를 확인해주세요."));
-	}
-
-}
