+++ src/main/java/com/goormitrip/goormitrip/user/controller/AuthController.java
@@ -1,0 +1,34 @@
+package com.goormitrip.goormitrip.user.controller;
+
+import com.goormitrip.goormitrip.user.dto.AuthRequest;
+import com.goormitrip.goormitrip.user.dto.AuthResponse;
+import com.goormitrip.goormitrip.user.dto.SignupRequest;
+import com.goormitrip.goormitrip.user.service.AuthService;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+
+@RestController
+@RequestMapping("/users")
+@RequiredArgsConstructor
+public class AuthController {
+
+    private final AuthService authService;
+
+    @PostMapping("/signup")
+    public ResponseEntity<AuthResponse> signup(@RequestBody @Valid SignupRequest request, Error error) {
+        // authService.signup(request);
+        return ResponseEntity.ok(authService.signup(request));
+    }
+
+    @PostMapping("/login")
+    public ResponseEntity<AuthResponse> login(@RequestBody @Valid AuthRequest request) {
+        return ResponseEntity.ok(authService.login(request));
+    }
+}
