+++ src/main/java/com/goormitrip/goormitrip/user/dto/AuthResponse.java
@@ -1,0 +1,15 @@
+package com.goormitrip.goormitrip.user.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthResponse {
+
+	private String token;
+	private String email;
+	private String role;
+}
