+++ src/main/java/com/goormitrip/goormitrip/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.goormitrip.goormitrip.user.repository;
+
+import com.goormitrip.goormitrip.user.domain.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    boolean existsByEmail(String email);
+
+    Optional<UserEntity> findByEmail(String email);
+}
