--- src/main/java/com/goormitrip/service/ReservationService.java
@@ -1,8 +1,0 @@
-package com.goormitrip.service;
-
-import java.time.LocalDate;
-import java.util.List;
-
-public interface ReservationService {
-    List<LocalDate> getAvailableDates(Long productId);
-}
